Move clustermesh.synced to common.synced
Move the 'synced' type to an exported 'Synced' in pkg/clustermesh/common
to allow for reuse in KVStoreMesh.

Since we are passing control of the StoppableWaitGroup for Services over
to the ServiceCache, it's necessary to expose this field on the struct
rather than helper methods as used for the other fields.